requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-telegram-bot==20.3
waitress==2.1.2
python-dotenv==1.0.0